.history/requirements_20251026191134.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
pydantic
pydantic-settings==2.1.0
python-dotenv
alembic